Feature chat

+++ backend/unithis/src/main/java/com/unithis/mapper/ChatRoomMapper.java
@@ -10,7 +10,7 @@ import com.unithis.model.ChatRoomRequest;
 
 @Mapper
 public interface ChatRoomMapper {
-	public int createChatRoom(ChatRoom newRoom);
+	public long createChatRoom(ChatRoom newRoom);
 	public List<ChatRoom> getJoinedRoomList(long userId);
 	public ChatRoom checkExistChatRoom(@Param("user1") long user1, @Param("user2") long user2);
 	public int updateReadTime(ChatRoomRequest chatroomReq);
